Refactor for Cyborg Module Rewriter tgui
