+++ tgui/packages/tgui/interfaces/CyborgModuleRewriter/EmptyPlaceholder.tsx
@@ -10,21 +10,13 @@ import { classes, pureComponentHooks } from 'common/react';
 import * as styles from './style';
 
 interface EmptyPlaceholderProps {
-  className?: string,
+  className?: string;
 }
 
-export const EmptyPlaceholder: SFC<EmptyPlaceholderProps> = props => {
-  const {
-    children,
-    className,
-  } = props;
-  const cn = classes([
-    styles.EmptyPlaceholder,
-    className,
-  ]);
-  return (
-    <div className={cn}>{children}</div>
-  );
+export const EmptyPlaceholder: SFC<EmptyPlaceholderProps> = (props) => {
+  const { children, className } = props;
+  const cn = classes([styles.EmptyPlaceholder, className]);
+  return <div className={cn}>{children}</div>;
 };
 
 EmptyPlaceholder.defaultHooks = pureComponentHooks;
