+++ tgui/packages/tgui/interfaces/CyborgModuleRewriter/action.ts
@@ -5,16 +5,18 @@
  * @license ISC
  */
 
-import { Action } from './type';
+import { Action } from './type/action';
 import type {
-  Act,
   ModuleActionPayload,
   MoveToolActionPayload,
   ResetModuleActionPayload,
   ToolActionPayload,
-} from './type';
+} from './type/action';
 
-const createAction = <T extends object>(action: Action) => (act: Act, payload: T) => act(action, payload);
+type Act = (action: string, payload?: object) => void;
+const createAction
+  = <T extends object>(action: Action) => (act: Act, payload: T) =>
+    act(action, payload);
 
 export const ejectModule = createAction<ModuleActionPayload>(Action.EjectModule);
 export const moveTool = createAction<MoveToolActionPayload>(Action.MoveTool);
