+++ tgui/packages/tgui/styles/interfaces/CyborgModuleRewriter.scss
@@ -13,11 +13,4 @@
   &-empty-placeholder {
     color: $color-label;
   }
-
-  &-module-view {
-    &__tool-label {
-      margin-left: 0.5em;
-      color: $color-label;
-    }
-  }
 }
